+++ base/src/test/java/org/perfmon4j/util/mbean/TestExampleMBean.java
@@ -1,5 +1,5 @@
 package org.perfmon4j.util.mbean;
 
 public interface TestExampleMBean {
-	public int getNextValue();
+	public long getNextValue();
 }
